Avoid double htmlspecialchars() in link ViewHelper
The fluid tag based viewhelper use already the htmlspecialchars()
function to escape special characters. Doing this in the dfgviewer xpath
viewhelper another time will result in rare cases in broken links.

e.g. &amp; is getting &amp;amp;

This patch adds an attribute "htmlspecialchars" which is TRUE by
default. Only in case of the link viewhelper, we set it to FALSE.